feynman-agent/requirements.txt
# Core dependencies for Feynman Agent
anthropic>=0.75.0
google-generativeai>=0.8.5
python-dotenv>=1.2.1
requests>=2.32.5

# Optional development dependencies
# Uncomment if you want to use these tools:
# pytest>=8.0.0
# pytest-mock>=3.12.0
# black>=24.0.0
# flake8>=7.0.0
# mypy>=1.8.0

